requirements.txt
feedparser
gitpython
heroku3
motor
git+https://github.com/KurimuzonAkuma/pyrogram@dev
python-dotenv
PyroMod==1.5
speedtest-cli
TgCrypto
httpx
